requirements.txt

Flask==2.3.3
google-cloud-firestore==2.11.1
google-generativeai==0.3.0
python-dotenv==1.0.0
Flask-Cors==4.0.0 
firebase-admin==5.2.0
numpy==1.26.4 
